Biblioteka_2.0/requirements.txt
flask
flask-sqlalchemy
flask-migrate
ipython
python-dotenv
flask-shell-ipython
